Added TODO, improved endpoint documentation

+++ src/users/users.controller.ts
@@ -7,12 +7,14 @@ import {
   Param,
   Delete,
   Query,
+  NotFoundException,
 } from '@nestjs/common';
 import { UsersService } from './users.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import {
   ApiCreatedResponse,
+  ApiNotFoundResponse,
   ApiOkResponse,
   ApiOperation,
   ApiTags,
@@ -26,37 +28,63 @@ export class UsersController {
 
   @Post()
   @ApiOperation({ summary: 'Create new user' })
-  @ApiOkResponse({ type: UserDto })
+  @ApiOkResponse({ description: 'Operation succeeded', type: UserDto })
   @ApiCreatedResponse({ description: 'Successfully created entity' })
   create(@Body() createUserDto: CreateUserDto) {
     return this.usersService.create(createUserDto);
   }
 
   @Get()
+  //TODO: Add documentation to query params, list structure
   @ApiOperation({ summary: 'Get user list' })
-  @ApiOkResponse({ type: UserDto, isArray: true })
+  @ApiOkResponse({
+    description: 'Operation succeeded',
+    type: UserDto,
+    isArray: true,
+  })
   findAll(@Query('crudQuery') crudQuery: string) {
     return this.usersService.findMany({ crudQuery });
   }
 
   @Get(':id')
   @ApiOperation({ summary: 'Get information for a user' })
-  @ApiOkResponse({ type: UserDto })
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+  @ApiOkResponse({ description: 'Operation succeeded', type: UserDto })
+  @ApiNotFoundResponse({ description: 'User not found' })
+  async findOne(@Param('id') id: string) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
+    return user;
   }
 
   @Patch(':id')
   @ApiOperation({ summary: 'Update a user' })
-  @ApiOkResponse({ type: UserDto })
-  update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+  @ApiOkResponse({ description: 'Operation succeeded', type: UserDto })
+  @ApiNotFoundResponse({ description: 'User not found' })
+  async update(@Param('id') id: string, @Body() updateUserDto: UpdateUserDto) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     return this.usersService.update(id, updateUserDto);
   }
 
   @Delete(':id')
   @ApiOperation({ summary: 'Delete a user' })
-  @ApiOkResponse({})
-  remove(@Param('id') id: string) {
+  @ApiOkResponse({ description: 'Operation succeeded' })
+  @ApiNotFoundResponse({ description: 'User not found' })
+  async remove(@Param('id') id: string) {
+    const user = await this.usersService.findOne(id);
+
+    if (!user) {
+      throw new NotFoundException('User not found');
+    }
+
     return this.usersService.remove(id, null);
   }
 }
